requirements.txt
python-binance
tabulate
